now it sorts boosts and powerhistory
for some unfathomable reason, when i ran the database off atlas it was messing up the order of these subarrays. that's what was causing the rankings to come out all weird. it never did this when i ran the database locally. no earthly idea why it happens. but now i've got a nice brute force fix in there